+++ logger/src/main/java/com/jeanbarrossilva/h2o/logger/InMemoryLogger.kt
@@ -1,7 +1,7 @@
 package com.jeanbarrossilva.h2o.logger
 
-import com.jeanbarrossilva.h2o.logger.intake.Intake
-import com.jeanbarrossilva.h2o.logger.intake.IntakeLog
+import com.jeanbarrossilva.h2o.model.intake.Intake
+import com.jeanbarrossilva.h2o.model.intake.IntakeLog
 import com.jeanbarrossilva.h2o.logger.time.MomentProvider
 
 data class InMemoryLogger(private val momentProvider: MomentProvider): Logger {
