--- logger/src/main/java/com/jeanbarrossilva/h2o/logger/extensions/intake/Intake.kt
@@ -1,6 +1,0 @@
-package com.jeanbarrossilva.h2o.logger.extensions.intake
-
-import com.jeanbarrossilva.h2o.logger.intake.Intake
-
-val Intake?.orZero
-    get() = this ?: Intake.zero
