+++ model/src/main/java/com/jeanbarrossilva/h2o/model/extensions/Intake.kt
@@ -1,0 +1,6 @@
+package com.jeanbarrossilva.h2o.model.extensions
+
+import com.jeanbarrossilva.h2o.model.intake.Intake
+
+val Intake?.orZero
+    get() = this ?: Intake.zero
