+++ settings.gradle
@@ -3,6 +3,6 @@ include ':app'
 include ':logger'
 include ':calculator'
 include ':preferences'
-include ':drinker'
 include ':ui'
 include ':ui-today'
+include ':model'
