+++ ui-today/src/main/java/com/jeanbarrossilva/h2o/ui/today/Today.kt
@@ -23,8 +23,8 @@ import androidx.compose.ui.layout.onGloballyPositioned
 import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.jeanbarrossilva.h2o.logger.intake.Intake
-import com.jeanbarrossilva.h2o.logger.intake.IntakeLog
+import com.jeanbarrossilva.h2o.model.intake.Intake
+import com.jeanbarrossilva.h2o.model.intake.IntakeLog
 import com.jeanbarrossilva.h2o.ui.component.Background
 import com.jeanbarrossilva.h2o.ui.component.FloatingActionButton
 import com.jeanbarrossilva.h2o.ui.environment.Spacing
